refactor(dag): decompose monolithic dag.rs into staging, nodes, graph, and builder
